chore: ignores README-only changes for actions triggers (#199)
While working on #187, I noticed that we "waste" a bit of Actions time rebuilding, linting, and running CodeQL against PRs that only introduce changes to `README.md`. Since we aren't using a markdown linter like `remark`, these Actions are all going to be a no-op.

So, this PR disables these actions on commits that only change `README.md`. If this is useful, we may want to:

* expand the set of files that we don't run changes against
* push this change upstream to other react projects we use, particularly the react starters that Teach LA uses